TPIII/Hit2/master/requirements.txt
google-cloud-storage
numpy
opencv-python-headless
docker
Pillow
